Change size/length types, fix warnings on 32-bit.

+++ src/tests/test_inet.c
@@ -71,8 +71,8 @@ int rrr_test_inet (void) {
 	int result = 0;
 
 	if (target_len != sizeof(in)) {
-		TEST_MSG("IPv4 mapped IPv6 to IPv4 convesion failure: Size mismatch %u vs %lu\n",
-			target_len, sizeof(in));
+		TEST_MSG("IPv4 mapped IPv6 to IPv4 convesion failure: Size mismatch %u vs %llu\n",
+			target_len, (long long unsigned) sizeof(in));
 		result |= 1;
 	}
